Rename snap_max_benefit variable to snap_max_allotment
