--- openfisca_us/variables/usda/snap/snap_max_benefit.py
@@ -1,16 +1,0 @@
-from openfisca_us.model_api import *
-
-
-class snap_max_benefit(Variable):
-    value_type = float
-    entity = SPMUnit
-    definition_period = YEAR
-    documentation = "Maximum benefit for SPM unit, based on the state group and household size."
-    label = "SNAP maximum benefit"
-    unit = USD
-
-    def formula(spm_unit, period, parameters):
-        snap_max_benefits = parameters(period).usda.snap.max_allotment.main
-        state_group = spm_unit.household("state_group_str", period)
-        household_size = spm_unit("spm_unit_size", period)
-        return snap_max_benefits[state_group][household_size] * 12
